+++ tests/testthat/test-tables.R
@@ -1,42 +1,42 @@
-# one_way() ####
+# tbl_one() ####
 test_that("one way generates", {
-  expect_no_condition(one_way(mtcars, ~gear))
+  expect_no_condition(tbl_one(mtcars, ~gear))
 })
 
 test_that("one way generates with digits", {
-  expect_no_condition(one_way(mtcars, ~gear, digits = 2))
+  expect_no_condition(tbl_one(mtcars, ~gear, digits = 2))
 })
 
 test_that("one way generates with caption", {
-  expect_no_condition(one_way(mtcars, ~gear, caption = "hehe"))
+  expect_no_condition(tbl_one(mtcars, ~gear, caption = "hehe"))
 })
 
 test_that("one way fails with mispelling", {
-  expect_error(one_way(mtcars, ~Gear))
+  expect_error(tbl_one(mtcars, ~Gear))
 })
 
 
-# two_way() ####
+# tbl_two() ####
 test_that("two way generates", {
-  expect_no_condition(two_way(mtcars, cyl~gear))
+  expect_no_condition(tbl_two(mtcars, cyl~gear))
 })
 
 test_that("two way generates with caption", {
-  expect_no_condition(two_way(mtcars, cyl~gear, caption = "hehe"))
+  expect_no_condition(tbl_two(mtcars, cyl~gear, caption = "hehe"))
 })
 
 test_that("two way generates with digits", {
-  expect_no_condition(two_way(mtcars, cyl~gear, digits = 2))
+  expect_no_condition(tbl_two(mtcars, cyl~gear, digits = 2))
 })
 
 test_that("two way generates with row percents", {
-  expect_no_condition(two_way(mtcars, cyl~gear, row_pct = TRUE))
+  expect_no_condition(tbl_two(mtcars, cyl~gear, row_pct = TRUE))
 })
 
 test_that("two way fails with wrong row percent argument", {
-  expect_error(two_way(mtcars, cyl~gear, row_pct = "yes"))
+  expect_error(tbl_two(mtcars, cyl~gear, row_pct = "yes"))
 })
 
 test_that("two way fails with mispelling", {
-  expect_error(two_way(mtcars, Cyl~Gear))
+  expect_error(tbl_two(mtcars, Cyl~Gear))
 })