+++ NEWS.md
@@ -1,6 +1,10 @@
 # gvsu215 (development version)
 
-* `two_way()` output cleaned up
+## Breaking Changes
+
+* Two functions have been renamed. `one_way()` is now `tbl_one()` and `two_way()` is now `tbl_two()`. The old functions are not available.
+
+* `two_way()` (`tbl_two()`) output cleaned up
 * `plot_box()` and `plot_hist()` will now not do scientific notation between certain ranges of numbers
 * `corr()` now shows the observations used (since this is not always just a sum of n1 and n2 missing)
 * added confidence interval to `infer_paired()`